# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: InvoiceOps API
  version: 1.0.0
  description: Contract for bilingual invoice and payment ticket triage.
servers:
  - url: http://localhost:8000
...
paths:
  /api/v1/auth/token:
    post:
      tags: [Authentication]
      summary: Issue a local PoC JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string, minLength: 1}
                password: {type: string, minLength: 1}
      responses:
        '200':
          description: Bearer token
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type, role]
                properties:
                  access_token: {type: string}
                  token_type: {type: string, const: bearer}
                  role: {type: string, enum: [admin, reviewer, observer]}
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/classifications:
    post:
      tags: [Classification]
      summary: Classify one supplier ticket
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassificationRequest'
            example:
              request_id: 9db8b9b1-7d2c-4d0e-8d4e-4f936d177a01
              source: email
              text: 请输入工单文本。
              taxonomy_version: invoiceops-v1
              metadata:
                channel: supplier_portal
      responses:
        '200':
          description: Classification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /api/v1/batches:
    post:
      tags: [Batches]
      summary: Submit a UTF-8 CSV batch
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: UTF-8 CSV matching docs/api/batch-template.csv.
      responses:
        '202':
          description: Batch accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchAccepted'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
...
  /api/v1/reviews:
    get:
      tags: [Reviews]
      summary: List tickets awaiting human review
      security: [{bearerAuth: []}]
      parameters:
        - in: query
          name: risk
          schema: {type: string, enum: [low, medium, high]}
        - in: query
          name: label_code
          schema: {type: string}
        - in: query
          name: waiting_min_seconds
          schema: {type: integer, minimum: 0}
        - in: query
          name: limit
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
      responses:
        '200':
          description: Review queue
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/ReviewItem'}
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/reviews/{ticket_id}/decisions:
    post:
      tags: [Reviews]
      summary: Append a human review decision
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewDecisionRequest'
      responses:
        '201':
          description: Decision appended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDecisionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Caller-generated key for safe retries.
      schema: {type: string, minLength: 1, maxLength: 255}
...
    TicketId:
      in: path
      name: ticket_id
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    IdempotencyConflict:
      description: Idempotency key was reused with a different request body
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Forbidden:
      description: Authenticated caller lacks permission
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, trace_id]
      properties:
        code: {type: string}
        message: {type: string}
        trace_id: {type: string}
        details: {type: object}
    ClassificationRequest:
      type: object
      additionalProperties: false
      required: [request_id, source, text, taxonomy_version]
      properties:
        request_id: {type: string, minLength: 1, maxLength: 128}
        source: {type: string, enum: [email, portal, api, manual]}
        text: {type: string, minLength: 1, maxLength: 5000}
        taxonomy_version: {type: string, const: invoiceops-v1}
        metadata:
          type: object
          additionalProperties: false
          properties:
            channel: {type: string, enum: [supplier_portal, email, internal]}
            locale: {type: string, enum: [zh, en, mixed]}
    ClassificationResponse:
      type: object
      required: [request_id, decision, language, predictions, route, model_version, threshold_version, taxonomy_version, trace_id]
      properties:
        request_id: {type: string}
        decision: {type: string, enum: [classified, needs_review]}
        reason_codes:
          type: array
          items: {type: string}
        language: {type: string, enum: [zh, en, mixed, unknown]}
        predictions:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [label_code, score]
            properties:
              label_code: {type: string}
              score: {type: number, minimum: 0, maximum: 1}
        route:
          type: object
          required: [primary, collaborators, version]
          properties:
            primary: {type: string}
            collaborators:
              type: array
              items: {type: string}
            version: {type: string}
        model_version: {type: string}
        threshold_version: {type: string}
        taxonomy_version: {type: string}
        inference_ms: {type: number, minimum: 0}
        trace_id: {type: string}
    BatchAccepted:
      type: object
      required: [batch_id, status, trace_id]
      properties:
        batch_id: {type: string, format: uuid}
        status: {type: string, enum: [queued, processing, completed, failed]}
        trace_id: {type: string}
...
    ReviewItem:
      type: object
      required: [ticket_id, request_id, risk, decision, predictions, reason_codes, primary_queue, collaborator_queues, waiting_seconds, overdue, review_sla_minutes]
      properties:
        ticket_id: {type: string, format: uuid}
        request_id: {type: string}
        risk: {type: string, enum: [low, medium, high]}
        decision: {type: string, const: needs_review}
        predictions: {type: array, items: {$ref: '#/components/schemas/LabelScore'}}
        reason_codes: {type: array, items: {type: string}}
        primary_queue: {type: string}
        collaborator_queues: {type: array, items: {type: string}}
        sanitized_text: {type: string}
        waiting_seconds: {type: integer, minimum: 0}
        overdue: {type: boolean}
        review_sla_minutes: {type: integer, minimum: 1}
...
    LabelScore:
      type: object
      required: [label_code, score]
      properties:
        label_code: {type: string}
        score: {type: number, minimum: 0, maximum: 1}
    ReviewDecisionRequest:
      type: object
      additionalProperties: false
      required: [labels, primary_queue]
      properties:
        labels: {type: array, minItems: 1, items: {type: string}}
        primary_queue: {type: string}
        note: {type: string, maxLength: 2000}
    ReviewDecisionResponse:
      type: object
      required: [ticket_id, revision, labels, reviewer_id, trace_id]
      properties:
        ticket_id: {type: string, format: uuid}
        revision: {type: integer, minimum: 1}
        labels: {type: array, items: {type: string}}
        reviewer_id: {type: string}
        trace_id: {type: string}
...